[close #48] Fix encoding issues
Problem
-------

Pandas decodes data with Latin-1
which lead to problems with german umlauts.

Solution
--------

* Data is now encoded back to Latin-1
  to minimize en/decoding errors.
* Add flag `--latin1` and `-l` to demarcate that the source files are
  Latin-1 encoded.

Explanation
-----------

We have two expected cases, keeping in mind
that utf-8 output is desired.

1. Input files are utf-8 encoded: decoding with subsequent encoding
   in Latin-1 will moth likely leave the original utf-8 encoding intact.
2. Input files are Latin-1 or Windows-1252 encoded: pandas decodes these
   files correctly, which means that we are working with utf-8 strings
   in the pandas object. This can then safely encoded to utf-8.

We also have one case that cannot be controlled for:

If the input is encoded in neither one of the expected encodings,
Latin-1 de and encoding will probably also leave this encoding intact.
Since, at the moment, pandas StataReader decoding is
hard coded to Latin-1, one Latin-1 decode, encode circle is necessary.
Offering more flexibility in the input encoding would necessitate
rereading the output in the correct encoding and writing it back to
utf-8 which just complicates the process and slows it down.
Input should generally be kept utf-8 encoded.